+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/ContributionProvidersBoxTestGenerated.java
@@ -248,6 +248,12 @@ public class ContributionProvidersBoxTestGenerated extends AbstractContributionP
   @TestDataPath("$PROJECT_ROOT")
   public class Contributionproviders {
     @Test
+    @TestMetadata("abstractedAssistedFactoryIsAccessibleExternally.kt")
+    public void testAbstractedAssistedFactoryIsAccessibleExternally() {
+      runTest("compiler-tests/src/test/data/box/aggregation/contributionproviders/abstractedAssistedFactoryIsAccessibleExternally.kt");
+    }
+
+    @Test
     public void testAllFilesPresentInContributionproviders() {
       KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("compiler-tests/src/test/data/box/aggregation/contributionproviders"), Pattern.compile("^(.+)\\.kt$"), null, true);
     }